requirements.txt
python-telegram-bot==11.1.0
peewee==3.9.6
